# api/src/static/api/api.yml
openapi: 3.0.0
info:
  title: User Service Database API
  description: 'Api rest to register users in a database for applications. Go to [Home page](http://localhost:5000/) for more information.'
  version: 1.0.0
  termsOfService: 'http://localhost:5000'
  contact:
    email: rodrjuniorsantos@gmail.com
servers:
  - url: 'http://localhost:5000'
    description: 'Primary server, e.g. Main (production) server'
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: api-token
paths:
...
  /v1/api/profiles/:
    post:
      summary: Create Profile Function.
      security:
        - APIKeyHeader: []
      tags:
        - Profile
      requestBody:
        description: Profile Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - career
                - birthDate
                - sex
                - telephone
              properties:
                career:
                  type: string
                birthDate:
                  type: string
                sex:
                  type: string
                telephone:
                  type: string
                married:
                  type: boolean
                cpf:
                  type: integer
                rg:
                  type: integer
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
                  
      responses:
        '200':
          description: Profile successfully registered
        '400':
          description: Error in request
        '401':
          description: Account in post not found
        '402':
          description: Missing data
    get:
      summary: Get all profiles Function
      security:
        - APIKeyHeader: []
      tags:
        - Profile
      responses:
        '200':
          description: Returns all profiles
        '400':
          description: User not found
        '401':
          description: Permission denied
  /v1/api/profiles/{profile_id}:
    get:
      summary: Gets a profile by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The profile ID.

      responses:
        '200':
          description: Data this profile successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
    put:
      summary: Update A Profile.
      security:
        - APIKeyHeader: []
      tags:
        - Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The user ID.
      requestBody:
        description: Profile Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - career
                - birthDate
                - sex
                - telephone
              properties:
                career:
                  type: string
                birthDate:
                  type: string
                sex:
                  type: string
                telephone:
                  type: string
                married:
                  type: boolean
                cpf:
                  type: integer
                rg:
                  type: integer
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
      responses:
        '200':
          description: Profile successfully update
        '400':
          description: Error in request
        '401':
          description: Not found
        '402':
          description: Permission denied
        '403':
          description: Missing data
    delete:
      summary: Delete a profile by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The profile ID.
      responses:
        '200':
          description: Profile successfully deleted
        '400':
          description: Profile not found
        '401':
          description: Permission denied
...